pyproject.toml
[project]
name = "api-contract-debugger-env"
version = "1.0.0"
description = "OpenEnv environment for debugging broken API contract specifications"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
